POO - Introducao Metodos Retorno 2

+++ src/cauafelype/maratona/java/javacore/introducaometodos/dominio/calculadora.java
@@ -8,10 +8,10 @@ public class calculadora {
         System.out.println(15-10);
     }
 
-    public int multiplicaDoisNumeros(int num1, int num2){
+    public double multiplicaDoisNumeros(double num1, double num2){
         return num1*num2;
     }
     public double divideDoisNumeros(double num1, double num2){
-        return num1/num2;
+            return num1 / num2;
     }
 }
